Override current working directory for "file=" queries in list cmd
When specifying a package using "file=" syntax, the result differs depending
on whether the current working directory is inside the Go module.
If inside the module, the package includes syntax trees for all files in the package.
If outside the module, the package includes only syntax trees for the specific file.
We want the same behavior in either case, so set the directory to the one containing
the requested file to guarantee that the current working directory is in the module.